refactor; ServiceName and ServiceVersion moved to config package
